Private field.
With a dynamic property you can mark a field as "private". This means that if the entity you are looking is "private", the framework sets the widget in a blinded state, not editable.
You can define "when" a particular entity is "private", annotating a method with the annotation "@Private". These method takes the entity as parameter, and should return true or false.